+++ prisma/migrations/20231126044710_add_digital_order_id_to_database/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "Order" ADD COLUMN     "digital_order_id" TEXT;
+
+-- AlterTable
+ALTER TABLE "VerificationEmail" ALTER COLUMN "expiration_date" SET DEFAULT NOW() + interval '1 day';