+++ prisma/migrations/20231126083809_/migration.sql
@@ -1,0 +1,19 @@
+/*
+  Warnings:
+
+  - The primary key for the `Item` table will be changed. If it partially fails, the table could be left without primary key constraint.
+  - You are about to drop the column `item_id` on the `Item` table. All the data in the column will be lost.
+  - Added the required column `model_id` to the `Item` table without a default value. This is not possible if the table is not empty.
+
+*/
+-- AlterTable
+ALTER TABLE "Item" DROP CONSTRAINT "Item_pkey",
+DROP COLUMN "item_id",
+ADD COLUMN     "model_id" TEXT NOT NULL,
+ADD CONSTRAINT "Item_pkey" PRIMARY KEY ("model_id", "order_id");
+
+-- AlterTable
+ALTER TABLE "VerificationEmail" ALTER COLUMN "expiration_date" SET DEFAULT NOW() + interval '1 day';
+
+-- AddForeignKey
+ALTER TABLE "Item" ADD CONSTRAINT "Item_model_id_fkey" FOREIGN KEY ("model_id") REFERENCES "Model"("id") ON DELETE RESTRICT ON UPDATE CASCADE;