+++ prisma/migrations/20231126085813_/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "ModelPromotion" ALTER COLUMN "discount" SET DATA TYPE DOUBLE PRECISION;
+
+-- AlterTable
+ALTER TABLE "VerificationEmail" ALTER COLUMN "expiration_date" SET DEFAULT NOW() + interval '1 day';