+++ src/configs/env.ts
@@ -1,5 +1,5 @@
 import { config as configEnv } from 'dotenv';
-import { str, num, cleanEnv } from 'envalid';
+import { str, num, cleanEnv, url } from 'envalid';
 
 configEnv();
 
@@ -15,7 +15,14 @@ export const envs = cleanEnv(process.env, {
     SMTP_SERVER: str(),
     SMTP_PORT: num(),
     SMTP_USER: str(),
-    SMTP_PASSWORD: str()
+    SMTP_PASSWORD: str(),
+    CHECKOUT_ENVIRONMENT: str(),
+    PAYPAL_LIVE_ENDPOINT: url(),
+    PAYPAL_SANDBOX_ENDPOINT: url(),
+    PAYPAL_CLIENT_ID: str(),
+    PAYPAL_CLIENT_SECRET: str(),
+    MINIO_URL: str()
 });
 
 export const CORS_WHITE_LIST = envs.CORS_WHITE_LIST.split(',');
+export const PAYPAL_ENDPOINT = envs.CHECKOUT_ENVIRONMENT === 'live' ? envs.PAYPAL_LIVE_ENDPOINT : envs.PAYPAL_SANDBOX_ENDPOINT;