+++ src/dtos/in/index.ts
@@ -4,6 +4,8 @@
 
 export * from './addCart.dto';
 export * from './auth.dto';
+export * from './order.dto';
+export * from './paypal.dto';
 export * from './register.dto';
 export * from './updateDefaultModel.dto';
 export * from './updateUserModel.dto';