+++ src/handlers/index.ts
@@ -5,5 +5,7 @@
 export * from './auth.handler';
 export * from './cart.handler';
 export * from './defaultModel.handler';
+export * from './checkout.handler';
+export * from './home.handler';
 export * from './user.handler';
 export * from './userModel.handler';