+++ src/types/handler.d.ts
@@ -1,2 +1,2 @@
 type Result<T> = Promise<T | void>;
-type HandlerTag = 'User' | 'Auth' | 'Cart' | 'DefaultModel' | 'UserModel';
+type HandlerTag = 'User' | 'Auth' | 'Cart' | 'DefaultModel' | 'UserModel' | 'Checkout' | 'Home';